Fixing Linting errors
