+++ front-end/src/App.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import './App.css';
+import LandingPage from './components/landingpage/landingpage';
+
+const App = () => (
+  <div className="App">
+    <LandingPage />
+  </div>
+);
+
+export default App;
