+++ front-end/src/components/landingpage/buyNowButton.jsx
