--- front-end/src/components/landingpage/landingpage.js
@@ -1,10 +1,0 @@
-import React from 'react';
-
-const LandingPage = () => (
-  <div>
-    <div>
-    </div>
-  </div>
-);
-
-export default LandingPage;
