+++ front-end/src/components/landingpage/landingpage.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import BuyNowButton from './buyNowButton';
+import LoginButtons from './signInSignUp';
+
+const LandingPage = () => (
+  <div>
+    <LoginButtons />
+    <BuyNowButton />
+  </div>
+);
+
+export default LandingPage;
