+++ front-end/src/components/landingpage/mainImage.jsx
@@ -1,0 +1,1 @@
+import React, { Component } from 'react';
