--- front-end/src/components/landingpage/signInSignOut.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import { ButtonToolbar, Button } from 'react-bootstrap';
-
-const LoginButtons = () => {
-  return (
-    <div>
-      <ButtonToolbar>
-        <Button>Sign Up</Button>
-        <Button>Sign In</Button>
-      </ButtonToolbar>
-    </div>
-  )
-};
-
-export default LoginButtons;
